FIX: Get shaking duration for normalizing spectrum from magnitude, not Arias intensity.
There is a problem with computing the duration from the "event" (noise + signal) using Arias
intensity as the SNR decreases. The noise artificially lengthens the duration, so the resulting
normalized signal spectrum becomes artificially smaller. An estimate of the shaking duration
computed from earthquake magnitude is independent of the SNR and much more robust.