FIX: Remove duplicate code to fetch moment tensor when using STREC.
SubductionSelector.getSubductionType() will fetch the moment tensor
information if given the event id. Consequently, we do not need to get
the moment tensor information in STREC.from_event() because we pass it the
event id. If we do, then in the case that the moment tensor is not
available, we try to download it two times.

The proposed solution relies on SubductionSelector().getSubductionType() to
get the moment tensor information.