REFACT: Provide more user control over time of signal/noise split.
Switch using the P-phase arrival time to determine the signal/noise
split to a user specified list of P-phase pickers.

The P-phase arrival time is used to provide a user-specified window for
the P-phase pickers (also specified by the user). The user can specify
the method for combining the picks (median, mean, max_snr).

A warning is logged if the P-phase pick differs from the arrival time by
some user-speficied amount.